Add a suitable pyproject.toml file.
With PEP 621, Python finally has support for storing project metadata in a
standardized format.  This file can eventually replace the setup.py file
that limits PiSCAT to using setuptools as its build system.

The pyproject.toml introduced with this commit uses Hatch as a build
system, and configures a variety of quality-of-life tools such as ruff,
back, isort, and pyright.