Use black and isort to reformat all files.
Also, remove a broken test case for downloading and using the tutorials.  The
current test were only causing trouble (as in, creating a 'Tutorials' directory
two levels above where the test suite is being run, or sometimes failing with
no apparent reason and a weird non-helpful message).

If we really want to have tests for the tutorials, we better start from scratch.